requirements.txt
tgtg==0.17.0
pytz==2024.2
python-telegram-bot==21.6
